fix problem with spaces in references generated by msgmerge tool
when msgmerge tool is used to generate the po file from a pot file,
references might be merged into one line, separated by spaces. There
seems to be a bug in pofile library. This is a work-around for this. It
should go upstream. However, if this fix goes upstream, this code will
still work, so it is safe to push this into a now release.