add support for context in compile_po
there has been no way to destinguish between strings with or without
context. Now those cases are handled, too. The context (if available) is
now added in front of the key, separated by \x00. Also plural forms are now
available directly as a key, because both msgids are added to the key,
separated by \x01. Both separators are unlikely to be used in strings that
should be translated.