filter messages marked as fuzzy from catalog by default
translators mark messages as fuzzy, if there is a problem with the
translation. Those messages need review and are most likely wrong. In
production environments, messages marked as fuzzy should not be
deployed, so these messages should be filtered out. Implementations can
than use the fallback translation (which might be the untranslated
string).