fix issue with gt.ngettext and variable as third argument
when (what is most likely in all projects the case) gt.ngettext is
called with a variable instead of a literal as the third argument (the
argument that is used to determine wether to use plural or singular
form), it hadn't been detected as a gt call.

This commit also adds a failing spec for this issue that works with the
patch.