add fallback for language extracted from po header
if no language is provided in po file header, fall back to the source
filename as language. This requires the source file to only have the
language as its name, which might be problematic. Therefore a warning is
shown, which might be changed to fail when grunt is not called with
"--force" parameter in the future.