fix issues with ambiguous Language headers
po documentation allows the Language header to contain an abbreviated
version of the Language name.

Therefor example files simple.fr_FR.po and fr_CA.po have been added.
Both have "Language: fr\n" header set, which makes the target ambiguous.
The convention is to silently fall back to fr_FR in this case (only if
the filename contains the complete string, fr_FR, in any way) or to
fall back to the filename itself, if the previous fallback mechanism
did not work.