fix potential race-condition in compile_po
processing of po files might not in all cases finish in order. So it is not
enough to check for the "last" file to be processed, but the number of
processed files is important. The task is done, if all files have been
processed.

Unfortunately it's hard to test this behaviour. May be, some day a decent
test-case can be written. For now I don't have the time.