add support for statically evaluable arguments for gt calls
sometimes it is good to split up strings into multiple parts, to distribute
them over more than one line. This makes the code more readable, but of course
results in a syntax subtree, that can not be used.

This new feature provides support for cases, that only do static
transformations to the specific arguments to the gt calls. So if a specific
argument can be evaluated to a string without any other scope, the argument
is considered valid, now.