don't fail if dynamic flag for gettext is set
if developers tell us by using the 'dynamic' flag in gettext style comments,
that this gt call is used with dynamic content, create_pot task will not fail
any longer. A warning will still be shown. This might be removed in the
future, once it's clear, we really want this behaviour.