fail with warning if item can not be extracted
until now, a warning has been written to console, if an item can not be
extracted from a po file, because it has no modules assigned to them.

This is considered an error, from now on. This is possible, because pofile
lib is now able to detect obsolete strings more reliable and therefor there
should be no strings without module references read from a po file.