optionally allow additional parameters for all function calls
Allow for all function calls to have additional parameters. This can be
used to automatically call a format function to help implementing
translated template strings, where optional parameters are used to
provide values for some variables.

Some examples, how this can be used in real code, can be found in the
specs (`spec/fixtures/simpleModule.js`). This example shows, that the
strings are extracted into the po file just fine.