Fix: compile_po does not create modules with wrong usage of gt and gt.ngettext
This is basically a follow up of
762bf3ca0713142a8b47840587db7193ad6cd221, which introduced create_pot
task to fail with duplicate message ids. This should also happen for
compile_po, because this is what is used mostly during development.
Since the duplicate ids should be fixed during development, failing the
compile_po task should help not running in this situation.

This can net really be tested, because of the static structure of grunt
configuration. So no test cases were added.